Comparison of SAT250
ex1 is the sat 250 where the [Return to addcaluse() from propagate()] represented by [addclause()---->propagate()----->addclause()] is changed from bin_count = 2(roughly) to 0.6. So, the time diff noticed will be close to 2.4 s and that is what we observe. 3.92-2.4 = 1.5 roughly equal to 1.22. So the reduction in the time can be noted.